api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
fastapi
flatbuffers==25.2.10
h11==0.14.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.44.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
onnxruntime==1.21.0
opencv-python-headless==4.11.0.86
packaging==24.2
pillow
platformdirs==4.3.7
pooch==1.8.2
protobuf==6.30.2
pydantic==2.10.6
pydantic_core==2.27.2
PyMatting==1.1.13
python-multipart
referencing==0.36.2
rembg
requests==2.32.3
rpds-py==0.24.0
scikit-image==0.25.2
scipy==1.15.2
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.3
tifffile==2025.3.13
tqdm==4.67.1
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn
